requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
altair==4.2.0
    # via streamlit
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via
    #   jsonschema
    #   streamlit
backcall==0.2.0
    # via ipython
backports-zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.0
    # via nbconvert
blinker==1.4
    # via streamlit
cachetools==5.2.0
    # via streamlit
certifi==2022.6.15
    # via requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via streamlit
commonmark==0.9.1
    # via rich
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   validators
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via
    #   altair
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fastjsonschema==2.15.3
    # via nbformat
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via streamlit
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via streamlit
importlib-resources==5.8.0
    # via jsonschema
ipykernel==6.15.0
    # via
    #   ipywidgets
    #   notebook
    #   pydeck
ipython==8.4.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
ipywidgets==7.7.0
    # via pydeck
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   altair
    #   nbconvert
    #   notebook
    #   pydeck
jsonschema==4.6.0
    # via
    #   altair
    #   nbformat
jupyter-client==7.3.4
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.10.0
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==1.1.0
    # via ipywidgets
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclient==0.6.4
    # via nbconvert
nbconvert==6.5.0
    # via notebook
nbformat==5.4.0
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via widgetsnbextension
numpy==1.22.4
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==21.3
    # via
    #   ipykernel
    #   nbconvert
    #   streamlit
pandas==1.4.2
    # via
    #   altair
    #   streamlit
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.1.1
    # via
    #   -r requirements.in
    #   streamlit
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.29
    # via ipython
protobuf==3.20.1
    # via streamlit
psutil==5.9.1
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyarrow==8.0.0
    # via streamlit
pycparser==2.21
    # via cffi
pydeck==0.7.1
    # via streamlit
pygments==2.12.0
    # via
    #   ipython
    #   nbconvert
    #   rich
pympler==1.0.1
    # via streamlit
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
    #   streamlit
pytz==2022.1
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyzmq==23.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
requests==2.28.0
    # via streamlit
rich==12.4.4
    # via streamlit
semver==2.13.0
    # via streamlit
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
smmap==5.0.0
    # via gitdb
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.3.0
    # via ipython
streamlit==1.10.0
    # via -r requirements.in
terminado==0.15.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
toml==0.10.2
    # via streamlit
toolz==0.11.2
    # via altair
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   streamlit
    #   terminado
traitlets==5.3.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   pydeck
typing-extensions==4.2.0
    # via
    #   rich
    #   streamlit
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via streamlit
urllib3==1.26.9
    # via requests
validators==0.20.0
    # via streamlit
watchdog==2.1.9
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==3.6.0
    # via ipywidgets
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
